requirements.txt
flask==3.1.0
Pillow==11.1.0
google-genai==1.5.0
python-dotenv==1.0.1
gunicorn==23.0.0
